Small changes and rebase

+++ src/main/java/mekanism/common/util/MekanismUtils.java
@@ -762,7 +762,7 @@ public final class MekanismUtils {
     }
 
     public static void veinMineArea(IEnergyContainer energyContainer, Level world, BlockPos pos, ServerPlayer player, ItemStack stack, Item usedTool,
-          Object2IntMap<BlockPos> found, Function<Float, FloatingLong> destroyEnergyFunction, BiFunction<Double, BlockState, Double> distanceMultiplier, BlockState sourceState) {
+          Object2IntMap<BlockPos> found, Function<Float, FloatingLong> destroyEnergyFunction, BiFunction<Integer, BlockState, Double> distanceMultiplier, BlockState sourceState) {
         FloatingLong energyUsed = FloatingLong.ZERO;
         FloatingLong energyAvailable = energyContainer.getEnergy();
         //Subtract from our available energy the amount that we will require to break the target block
@@ -780,7 +780,7 @@ public final class MekanismUtils {
             if (hardness == -1) {
                 continue;
             }
-            FloatingLong destroyEnergy = destroyEnergyFunction.apply(hardness).multiply(distanceMultiplier.apply((double) foundEntry.getIntValue(), targetState));
+            FloatingLong destroyEnergy = destroyEnergyFunction.apply(hardness).multiply(distanceMultiplier.apply(foundEntry.getIntValue(), targetState));
             if (energyUsed.add(destroyEnergy).greaterThan(energyAvailable)) {
                 //If we don't have energy to break the block continue
                 //Note: We do not break as given the energy scales with hardness, so it is possible we still have energy to break another block
